uint16_t plutot que uint8_t et changement appel a la fonction _delay_ms pour attendre_ms

+++ codeCommun/lib_c/delay.h
@@ -1,3 +1,3 @@
 #include "constantes.h"
 
-void attendre_ms(int millisecondes);
+void attendre_ms(uint16_t millisecondes);
